Move WorkspaceMembership to own file and add ORM todo
I've been considering switching to EF, because if this was in memory my
domain wouldn't relate different entities by Ids.
I'd have a User and Workspace in WorkspaceMembership, for example.
But how much would that gain me actually?
And, since relying on lazy loading rules out micro-Orms and no-sql is
there an irony in making that choice to 'decouple' domain and storage.
I don't know. I really don't.